Move document loading into AsnycMindmapLoaderTask
