+++ app/src/main/java/ch/benediktkoeppel/code/droidplane/MainActivity.java
@@ -85,6 +85,7 @@ public class MainActivity extends FragmentActivity {
         // if we already have a loaded mindmap, use this; otherwise load from the intent
         if (mindmap.isLoaded()) {
             horizontalMindmapView.setMindmap(mindmap);
+            horizontalMindmapView.onRootNodeLoaded();
 
         } else {
